listening-comp/pyproject.toml
[project]
name = "listening-comp"
version = "0.1.0"
description = "Language Listening Comprehension App"
requires-python = ">=3.9"
dependencies = [
    "gradio>=3.50.0",
    "yt-dlp>=2023.3.4",
    "whisper>=1.0",
    "opencv-python>=4.8.0",
    "easyocr>=1.7.0",
    "gtts>=2.3.2",
    "qdrant-client>=1.6.0",
    "openai>=0.28.0",
    "python-dotenv>=1.0.0",
    "youtube-transcript-api>=0.6.1",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "tesseract>=0.1.3"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
